pyproject.toml
[tool.poetry]
name = "peakguard-api"
version = "0.1.0"
description = "FastAPI service for synthetic energy data and GRU forecasts"
authors = ["You <you@example.com>"]
packages = [
  { include = "app" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
jinja2 = "^3.1.0"
pydantic = "^2.7.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
scikit-learn = "^1.5.0"
# Apple Silicon (arm64) TensorFlow
tensorflow-macos = "==2.15.0"
tensorflow-metal = ">=1.1.0"
keras = ">=3.2.0"
joblib = "^1.3.0"
python-multipart = "^0.0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
